+++ packages/qb/pqb/src/columns/operators.ts
@@ -1,13 +1,13 @@
-import { Query } from './query';
-import { addValue } from './sql/common';
-import { getRaw, isRaw, RawExpression } from './raw';
+import { Query } from '../query';
+import { addValue } from '../sql/common';
+import { getRaw, isRaw, RawExpression } from '../raw';
 
 type Fn<T> = (key: string, value: T, values: unknown[]) => string;
 
 export type Operator<T> = Fn<T> & { type: T };
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type Operators = Record<string, Operator<any>>;
+export type BaseOperators = Record<string, Operator<any>>;
 
 export const createOperator = <T>(fn: Fn<T>) => {
   return Object.assign(fn, { type: undefined as unknown as T });