+++ packages/schema-to-zod/CHANGELOG.md
@@ -1,5 +1,33 @@
 # orchid-orm-schema-to-zod
 
+## 0.2.23
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.9.17
+
+## 0.2.22
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.9.16
+
+## 0.2.21
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.9.15
+
+## 0.2.20
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.9.14
+
 ## 0.2.19
 
 ### Patch Changes